move test helpers to compiled directory
